Interface_V001
includes login - membership - main form - and cart with links attached
for proper flow.  Database is attached which allows for insertion of new
members.  Password and email need special formatting, login page needs
to be tested with the database to see if there is a member with such a
user name and password - main form and cart are blank at this time.